+++ library/Dlayer/DesignerTool/ContentManager/Html/SubTool/Delete/Tool.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Html delete sub tool
+ *
+ * @author Dean Blackborough <dean@g3d-development.com>
+ * @copyright G3D Development Limited
+ * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
+ */
+class Dlayer_DesignerTool_ContentManager_Html_SubTool_Delete_Tool extends
+    Dlayer_DesignerTool_ContentManager_Shared_Tool_Delete
+{
+    // No custom logic required
+}