+++ library/Dlayer/DesignerTool/ContentManager/Shared/Tool/Delete.php
@@ -1,0 +1,125 @@
+<?php
+
+/**
+ * Delete sub tool
+ *
+ * @author Dean Blackborough <dean@g3d-development.com>
+ * @copyright G3D Development Limited
+ * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
+ */
+class Dlayer_DesignerTool_ContentManager_Shared_Tool_Delete extends Dlayer_Tool_Content
+{
+    /**
+     * Check that the required params have been submitted, check the keys in the params array
+     *
+     * @param array $params
+     * @return boolean
+     */
+    protected function paramsExist(array $params)
+    {
+        $valid = false;
+        if (array_key_exists('confirm', $params) === true) {
+            $valid = true;
+        }
+
+        return $valid;
+    }
+
+    /**
+     * Check to ensure the posted params are of the correct type and optionally within range
+     *
+     * @param array $params
+     * @return boolean
+     */
+    protected function paramsValid(array $params)
+    {
+        $valid = false;
+        if (intval($params['confirm']) === 1) {
+            $valid = true;
+        }
+
+        return $valid;
+    }
+
+    /**
+     * Prepare the posted params, convert them to the required types and assign to the $this->params property
+     *
+     * @param array $params
+     * @param boolean $manual_tool Are the values to be assigned to $this->params or $this->params_auto
+     * @return void
+     */
+    protected function paramsAssign(array $params, $manual_tool = true)
+    {
+        $this->params = array(
+            'confirm' => true
+        );
+    }
+
+    /**
+     * Add a new content item or setting
+     *
+     * @return array|FALSE Ids for new environment vars or FALSE if the request failed
+     */
+    protected function add()
+    {
+        // Not required by tool
+    }
+
+    /**
+     * Edit a new content item or setting
+     *
+     * @return array|FALSE Ids for new environment vars or FALSE if the request failed
+     * @throws Exception
+     */
+    protected function edit()
+    {
+        $model = new Dlayer_DesignerTool_ContentManager_Text_SubTool_Delete_Model();
+
+        $deleted = $model->setDeleted($this->site_id, $this->page_id, $this->column_id, $this->content_id);
+
+        if ($deleted === true) {
+            $model->reorderColumnContent($this->site_id, $this->page_id, $this->column_id);
+        }
+
+        return true;
+    }
+
+    /**
+     * Make a structural change to the page
+     *
+     * @return array|FALSE An array of the environment vars to set or FALSE upon error
+     */
+    protected function structure()
+    {
+        // Not required by manual tool
+    }
+
+    /**
+     * Generate the return ids array
+     *
+     * @param integer|NULL Set content id or use the class property
+     * @return array
+     */
+    protected function returnIds($content_id = null)
+    {
+        return array(
+            array(
+                'type' => 'page_id',
+                'id' => $this->page_id,
+            )
+        );
+    }
+
+    /**
+     * Validate the instances param, need to see if it should exist first
+     *
+     * @param integer $site_id
+     * @param integer $content_id
+     *
+     * @return boolean
+     */
+    protected function validateInstances($site_id, $content_id)
+    {
+        // TODO: Implement validateInstances() method.
+    }
+}