+++ library/Dlayer/DesignerTool/ContentManager/Text/Ribbon.php
@@ -22,15 +22,17 @@ class Dlayer_DesignerTool_ContentManager_Text_Ribbon extends Dlayer_Ribbon_Conte
 
 		$this->contentData();
 		$this->previewData();
+		$this->instancesOfData();
 
 		return array(
 			'form' => new Dlayer_DesignerTool_ContentManager_Text_Form(
 			    $tool,
                 $this->content_data,
-				$this->instancesOfData(),
+				$this->instances_of,
                 array()
             ),
-            'preview' => $this->preview_data
+            'preview' => $this->preview_data,
+            'instances' => $this->instances_of
 		);
 	}
 
@@ -64,19 +66,21 @@ class Dlayer_DesignerTool_ContentManager_Text_Ribbon extends Dlayer_Ribbon_Conte
 	/**
 	 * Fetch the number of instances for the content item data
 	 *
-	 * @return integer
+	 * @return void
 	 */
 	protected function instancesOfData()
 	{
-		$instances = 0;
+	    if ($this->instances_of_fetched === false) {
 
-		if($this->tool['content_id'] !== NULL)
-		{
-			$model_text = new Dlayer_DesignerTool_ContentManager_Text_Model();
-			$instances = $model_text->instancesOfData($this->tool['site_id'], $this->tool['content_id']);
-		}
+	        $this->instances_of = 0;
 
-		return $instances;
+            if ($this->tool['content_id'] !== null) {
+                $model = new Dlayer_DesignerTool_ContentManager_Text_Model();
+                $this->instances_of = $model->instancesOfData($this->tool['site_id'], $this->tool['content_id']);
+            }
+
+            $this->instances_of_fetched = true;
+        }
 	}
 
     /**