+++ library/Dlayer/Ribbon/Handler/Content.php
@@ -234,6 +234,11 @@ class Dlayer_Ribbon_Handler_Content
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
                 break;
 
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Text_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
+                break;
+
             default:
                 $data = false;
                 break;
@@ -266,6 +271,11 @@ class Dlayer_Ribbon_Handler_Content
             case 'typography':
                 $ribbon_typography = new Dlayer_DesignerTool_ContentManager_Heading_SubTool_Typography_Ribbon();
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
+                break;
+
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Heading_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
                 break;
 
             default:
@@ -302,6 +312,11 @@ class Dlayer_Ribbon_Handler_Content
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
                 break;
 
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Html_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
+                break;
+
             default:
                 $data = false;
                 break;
@@ -334,6 +349,11 @@ class Dlayer_Ribbon_Handler_Content
             case 'typography':
                 $ribbon_typography = new Dlayer_DesignerTool_ContentManager_Jumbotron_SubTool_Typography_Ribbon();
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
+                break;
+
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Jumbotron_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
                 break;
 
             default:
@@ -370,6 +390,11 @@ class Dlayer_Ribbon_Handler_Content
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
                 break;
 
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Form_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
+                break;
+
             default:
                 $data = false;
                 break;
@@ -402,6 +427,11 @@ class Dlayer_Ribbon_Handler_Content
             case 'typography':
                 $ribbon_typography = new Dlayer_DesignerTool_ContentManager_Image_SubTool_Typography_Ribbon();
                 $data = $ribbon_typography->viewData($this->toolParams($tool));
+                break;
+
+            case 'delete':
+                $ribbon_delete = new Dlayer_DesignerTool_ContentManager_Image_SubTool_Delete_Ribbon();
+                $data = $ribbon_delete->viewData($this->toolParams($tool));
                 break;
 
             default: